refactor(llm): split router/mod.rs monolith into focused submodules
Split the 4140-line router/mod.rs (4 scattered `impl RouterProvider`
blocks plus inline cache and config structs) into focused submodules:

- config.rs        RouterStrategy and the three *RouterConfig structs
- embed_cache.rs   BanditEmbedCache, TurnEmbedCache
- builder.rs       constructor, with_* builders, save_*, *_stats
- select.rs        provider ordering/selection, embed cache, ASI updates
- chat.rs          bandit_chat, cascade_chat(_stream), stream helpers
- provider_impl.rs impl LlmProvider for RouterProvider
- tests.rs         unit tests

mod.rs (now ~170 lines) keeps the module docs, declarations and
re-exports, the RouterProvider struct, the shared blocking_load helper,
and the ASI rate-limit statics. Cross-submodule helper methods are
bumped to pub(crate); the public API is unchanged — RouterStrategy, the
config types, and RouterProvider remain re-exported from router.

Closes #4875
